Move PractitionerAddressTest module
