+++ lib/us_core_test_kit/custom_groups/v7.0.0/practitioner_address_test.rb
@@ -1,4 +1,4 @@
-require_relative '../../practitioner_address_test'
+require_relative '../practitioner_address_test'
 
 module USCoreTestKit
   module USCoreV700
